requirements.txt
streamlit
python-docx
PyPDF2
openai
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
